+++ src/MyLibrary/Presenters/AddBookPresenter.cs
@@ -92,7 +92,6 @@ namespace MyLibrary.Presenters
         /// <param name="book"></param>
         public void Prefill(Book book)
         {
-            // clear fields
             this._view.TitleFieldText = "";
             this._view.LongTitleFieldText = "";
             this._view.IsbnFieldText = "";
@@ -112,8 +111,6 @@ namespace MyLibrary.Presenters
             this._view.UncheckAllAuthors();
             this._view.UncheckAllTags();
 
-            // now populate
-            // text/number input fields
             this._view.TitleFieldText = book.Title;
             this._view.LongTitleFieldText = book.TitleLong;
             this._view.IsbnFieldText = book.Isbn;
@@ -122,14 +119,18 @@ namespace MyLibrary.Presenters
             this._view.PlaceOfPublicationFieldText = book.PlaceOfPublication;
             this._view.PagesFieldText = book.Pages.ToString();
             this._view.LanguageFieldText = book.Language;
-            // publisher
+
             if (!this._allPublishers.Contains(book.Publisher.Name))
             {
                 this._allPublishers.Add(book.Publisher.Name);
             }
             FilterPublishers(null, null);
             this._view.SetPublisher(book.Publisher, true);
-            // authors
+
+            foreach (var key in this._allAuthors.Keys.ToList())
+            {
+                this._allAuthors[key] = false;
+            }
             foreach (var author in book.Authors)
             {
                 if (!this._allAuthors.ContainsKey(author.GetFullNameLastNameCommaFirstName()))
@@ -146,11 +147,9 @@ namespace MyLibrary.Presenters
 
         public void FilterTags(object sender, EventArgs args)
         {
-            // grab the filter
             const RegexOptions REGEX_OPTIONS = RegexOptions.IgnoreCase | RegexOptions.ExplicitCapture;
             Regex filterPattern = new Regex(this._view.FilterTagsFieldEntry, REGEX_OPTIONS);
 
-            // perform filtering
             Dictionary<string, bool> filteredTags = new Dictionary<string, bool>();
             foreach (var kvp in this._allTags)
             {
@@ -160,17 +159,14 @@ namespace MyLibrary.Presenters
                 }
             }
 
-            // update the view
             this._view.AddTags(filteredTags);
         }//FilterTags
 
         public void FilterAuthors(object sender, EventArgs args)
         {
-            // grab the filter
             const RegexOptions REGEX_OPTIONS = RegexOptions.IgnoreCase | RegexOptions.ExplicitCapture;
             Regex filterPattern = new Regex(this._view.FilterAuthorsFieldEntry, REGEX_OPTIONS);
 
-            // perform filtering
             Dictionary<string, bool> filteredAuthors = new Dictionary<string, bool>();
             foreach (var kvp in this._allAuthors)
             {
@@ -180,17 +176,14 @@ namespace MyLibrary.Presenters
                 }
             }
 
-            // update the view
             this._view.AddAuthors(filteredAuthors);
         }//FilterAuthors
 
         public void FilterPublishers(object sender, EventArgs args)
         {
-            // grab the filter
             const RegexOptions REGEX_OPTIONS = RegexOptions.IgnoreCase | RegexOptions.ExplicitCapture;
             Regex filterPattern = new Regex(this._view.FilterPublishersFieldEntry, REGEX_OPTIONS);
 
-            // perform filtering
             List<string> filteredPublishers = new List<string>();
             foreach (var publisher in this._allPublishers)
             {
@@ -200,28 +193,26 @@ namespace MyLibrary.Presenters
                 }
             }
 
-            // update the view
             this._view.AddPublishers(filteredPublishers);
         }//FilterPublishers
 
         public async Task PopulateAuthorsList()
         {
-            // load all authors
             var allAuthors = await this._authorService.GetAll();
             foreach (var author in allAuthors)
             {
                 string authorName = author.GetFullNameLastNameCommaFirstName();
                 if (!this._allAuthors.ContainsKey(authorName))
+                {
                     this._allAuthors.Add(authorName, false);
+                }
             }
 
-            // perform filtering and update the view
             FilterAuthors(null, null);
         }//PopulateAuthorsList
         
         public async Task PopulateTagsList()
         {
-            // load all tags
             var allTags = await this._tagService.GetAll();
             foreach (var tag in allTags)
             {
@@ -232,13 +223,11 @@ namespace MyLibrary.Presenters
                 }
             }
 
-            // perform filtering and update the view
             FilterTags(null, null);
         }//PopulateTagsList
 
         public async Task PopulatePublishersList()
         {
-            // load all publishers
             var allPublishers = await this._publisherService.GetAll();
             foreach (var publisher in allPublishers)
             {
@@ -247,7 +236,6 @@ namespace MyLibrary.Presenters
                     this._allPublishers.Add(publisher.Name);
             }
 
-            // perform filtering and update the view
             FilterPublishers(null, null);
         }//PopulatePublishersList
 
@@ -433,7 +421,6 @@ namespace MyLibrary.Presenters
 
         public void HandleAddNewTagClicked(object sender, EventArgs args)
         {
-            //string newTag = ShowNewTagDialog();
             string newTag = this._view.ShowNewTagDialog();
             if (!string.IsNullOrWhiteSpace(newTag))
             {