+++ src/MyLibrary/Presenters/SearchByIsbnPresenter.cs
@@ -104,19 +104,7 @@ namespace MyLibrary.Presenters
             this._view.SearchButtonEnabled = false;
 
             string enteredIsbn = this._view.IsbnFieldText;
-            
-            // check if book with this ISBN already exists in database
-            if (await this._bookService.ExistsWithIsbnAsync(enteredIsbn))
-            {
-                // book already exists
-                // tell the user
-                this._view.ShowAlreadyExistsWithIsbnDialog(enteredIsbn);
 
-                Reset();
-
-                return;
-            }
-            
             Book book = null;
             try
             {